new settings, custom html

+++ app/Http/Controllers/Admin/SettingsController.php
@@ -23,27 +23,34 @@ class SettingsController extends Controller {
         $fields = Settings::getFieldsIfValid($request);
         $settings = Settings::all()->pluck('value', 'key')->toArray();
         foreach ($fields as $key => $value) {
-            if ($settings[$key] !== $fields[$key] || ($key === 'logo' && $value) || ($key === 'cover' && $value)) {
-                if (($key === 'logo' || $key === 'cover') && $value) {
-                    $path = "/public/img/$key";
-                    Storage::makeDirectory($path);
+            if (key_exists($key, $settings)) {
+                if ($settings[$key] !== $fields[$key] || ($key === 'logo' && $value) || ($key === 'cover' && $value)) {
+                    if (($key === 'logo' || $key === 'cover') && $value) {
+                        $path = "/public/img/$key";
+                        Storage::makeDirectory($path);
 
-                    // Original
-                    Storage::delete($path . '/' . $settings[$key]);
-                    $request->file($key)->storeAs($path, $value);
+                        // Original
+                        Storage::delete($path . '/' . $settings[$key]);
+                        $request->file($key)->storeAs($path, $value);
 
-                    if ($key === 'logo') {
-                        // Only png are supported
-                        $name = substr($value, 0, -4);
-                        $old_name = substr($settings['logo'], 0, -4);
-                        $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 72);
-                        $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 96);
-                        $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 128);
-                        $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 192);
-                        $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 256);
+                        if ($key === 'logo') {
+                            // Only png are supported
+                            $name = substr($value, 0, -4);
+                            $old_name = substr($settings['logo'], 0, -4);
+                            $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 72);
+                            $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 96);
+                            $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 128);
+                            $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 192);
+                            $this->convertAndStore($request->file('logo'), $path, $name, $old_name, 256);
+                        }
                     }
+                    Settings::where('key', $key)->update(['value' => $value]);
                 }
-                Settings::where('key', $key)->update(['value' => $value]);
+            } else {
+                $s = new Settings();
+                $s->key = $key;
+                $s->value = $value;
+                $s->save();
             }
         }
         Artisan::call('config:cache');
